SDK: Move SDK helpers into TestFramework
